SCM - 💄 extract input into a separate file
